Check if Scripts directory exists

+++ download.ps1
@@ -38,4 +38,6 @@ if ($env:PYVERSION -like "3.*") {
     Get-ChildItem -Include "*.pyc" -Recurse -Force | Remove-Item -Force
 }
 
-Remove-Item -Recurse -Force $targetdir\Scripts
+if (Test-Path $targetdir\Scripts) {
+    Remove-Item -Recurse -Force $targetdir\Scripts
+}
